+++ ui/mods/puppetmaster/live_game.js
@@ -66,12 +66,50 @@
     liveGameSelection(payload)
   }
 
-  var ping = function(location) {
+  var ping = function(armyIdx, location) {
     hdeck.view.sendOrder({
-      units: [commanderIds[armyIndex()]],
+      units: [commanderIds[armyIdx]],
       command: 'ping',
       location: location,
     })
+  }
+
+  var flare = function(loc, time) {
+    var puppet = {
+      "model": {
+        "filename": "/pa/units/land/land_mine/land_mine.papa", 
+      }, 
+      location: {
+        planet: loc.planet,
+        pos: loc.pos,
+        scale: 4,
+      },
+      material: {
+        shader: "pa_unit_ghost",
+        constants: {
+          GhostColor: [0,0,1,0],
+          BuildInfo: [0,10,0,0],
+        },
+        textures: {
+          Diffuse: "/pa/effects/diffuse_texture.papa"
+        }
+      },
+      "fx_offsets": [
+        {
+          "bone": "bone_root", 
+          "filename": "/pa/units/land/tank_armor/tank_armor_muzzle_flame.pfx", 
+          "offset": [ 0, 0, 0 ], 
+          "orientation": [ 0, 90, 0 ], 
+        }
+      ], 
+    }
+    var view = hdeck.view
+    view.puppet(puppet).then(function(r) {
+      puppet = r
+    })
+    setTimeout(function() {
+      view.unPuppet(puppet.id)
+    }, time)
   }
 
   // Spectator Announcement, including drop-pod effect
@@ -158,6 +196,8 @@
     return engineCall.apply(this, arguments);
   }
 
+  //api.getWorldView(0).clearPuppets()
+
   var pasteUnits = function(n) {
     if (!model.cheatAllowCreateUnit()) return
     if (n < 1) return
@@ -166,6 +206,10 @@
     var army_id = model.players()[armyIndex()].id
 
     hdeck.raycastWithPlanet(mouseX, mouseY).then(function(result) {
+      flare(result, 4100)
+
+      setTimeout(ping, 4000, armyIndex(), result)
+
       var drop = {
         army: army_id,
         what: dropPodSpec,
@@ -175,7 +219,6 @@
       pasteUnits3D(1, drop)
       drop.what = selectedUnit.spec
       setTimeout(pasteUnits3D, 5000, n, drop)
-      setTimeout(ping, 4000, result)
     })
 
     increment(n)